[services] Added StorageSyncJob

+++ core/src/main/kotlin/io/slychat/messenger/core/http/api/storage/StorageAsyncClient.kt
@@ -7,7 +7,7 @@ import nl.komponents.kovenant.Promise
 interface StorageAsyncClient {
     fun getQuota(userCredentials: UserCredentials): Promise<Quota, Exception>
 
-    fun getFileList(userCredentials: UserCredentials, sinceVersion: Int): Promise<FileListResponse, Exception>
+    fun getFileList(userCredentials: UserCredentials, sinceVersion: Long): Promise<FileListResponse, Exception>
 
     fun update(userCredentials: UserCredentials, request: UpdateRequest): Promise<UpdateResponse, Exception>
 }
